Progress dump of reorganization.
This doesn't make a great 'first real commit' for a portfolio project, but I had thrown together a project in a fairly unstructured manner and later decided I wanted to follow a better defined architecture for project composition. After looking around a bit I found one that's pretty onion-esque and seemed to really highlight the ability to mock any and all functionality out + have a high degree of abstraction on 'core logic'. This is the first committable step along that reorganization.